renamed jsx to js, moved PrimToConv into symmetry
