+++ lib/main.js
@@ -14,5 +14,6 @@ export {
 } from "./utils/formatting.jsx";
 
 export * as MatrixUtils from "./utils/matrixUtils.jsx";
+export * as FormulaUtils from "./utils/formulaUtils.jsx";
 
-export { getSymmetryInfo } from "./utils/symmetry.js";
+export { getSymmetryInfo, getPrimToConvMatrix } from "./utils/symmetry.js";
