+++ lib/utils/formulaUtils.js
@@ -1,5 +1,4 @@
 import { formatChemicalFormula } from "./formatting";
-import pt from "./pt.json"; // move to mc-r-lib?
 
 export function countNumberOfAtoms(formula) {
   // split on capital letters to get element+number strings
@@ -25,7 +24,7 @@ export function formatTitle(formulaStr, id, method) {
   );
 }
 
-export function calculateDensity(formula, volume) {
+export function calculateDensity(formula, pt, volume) {
   // currently uses scoped pt to determine densities.
   var elnum = formula.split(/(?=[A-Z])/);
   let totMz = 0;
