+++ lib/utils/symmetry.js
@@ -111,3 +111,61 @@ export function getSymmetryInfo(spgn) {
     bravais_lattice: bravais[1],
   };
 }
+
+export function getPrimToConvMatrix(spaceGroupSymbol) {
+  // Method to return the conv Matrix of a primitive cell
+  //  to the conventional standard cell
+
+  if (spaceGroupSymbol.includes("P")) {
+    return [
+      [1, 0, 0],
+      [0, 1, 0],
+      [0, 0, 1],
+    ];
+  } else if (spaceGroupSymbol.includes("I")) {
+    return [
+      [0, 1, 1],
+      [1, 0, 1],
+      [1, 1, 0],
+    ];
+  } else if (spaceGroupSymbol.includes("F")) {
+    return [
+      [-1, 1, 1],
+      [1, -1, 1],
+      [1, 1, -1],
+    ];
+  } else if (spaceGroupSymbol.includes("A")) {
+    return [
+      [1, 1, 0],
+      [-1, 1, 0],
+      [0, 0, 1],
+    ];
+  } else if (spaceGroupSymbol.includes("R")) {
+    return [
+      [-1, 1, 0],
+      [1, 1, -1],
+      [1, 1, 1],
+    ];
+  } else if (spaceGroupSymbol.includes("C")) {
+    if (spaceGroupSymbol in ["C2", "CmCc", "C2/m", "C2/c"]) {
+      return [
+        [1, -1, 0],
+        [1, 1, 0],
+        [0, 0, 1],
+      ];
+    } else {
+      return [
+        [1, -1, 0],
+        [1, 1, 0],
+        [0, 0, 1],
+      ];
+    }
+  } else {
+    console.log("Missing Space group data for:", spaceGroupSymbol);
+    return [
+      [1, 0, 0],
+      [0, 1, 0],
+      [0, 0, 1],
+    ];
+  }
+}
